backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

SQLAlchemy==2.0.32
psycopg[binary]==3.1.18

httpx==0.27.0
apscheduler==3.10.4

numpy==2.0.1
pandas==2.2.2
scikit-learn==1.5.1

yfinance==0.2.43
